+++ src/Control/Carrier/FossaApiClient/Internal/FossaAPIV1.hs
@@ -56,6 +56,7 @@ module Control.Carrier.FossaApiClient.Internal.FossaAPIV1 (
   getReleaseGroupReleases,
   updateReleaseGroupRelease,
   createReleaseGroup,
+  createReleaseGroupRelease,
 
   -- * Policy
   getPolicies,
@@ -101,6 +102,7 @@ import App.Types (
   ProjectMetadata (..),
   ProjectRevision (..),
   ReleaseGroupMetadata (releaseGroupName, releaseGroupRelease),
+  ReleaseGroupReleaseRevision,
   fullFileUploadsToCliLicenseScanType,
  )
 import App.Version (versionNumber)
@@ -1190,6 +1192,8 @@ getAttributionJson apiOpts ProjectRevision{..} = fossaReq $ do
             =: True
           <> "dependencyInfoOptions[]"
             =: packageDownloadUrl
+          -- Large reports can take over a minute to generate, so increase the timeout to 10 minutes
+          <> responseTimeoutSeconds 600
   orgId <- organizationId <$> getOrganization apiOpts
   response <- req GET (attributionEndpoint baseUrl orgId (Locator "custom" projectName (Just projectRevision)) ReportJson) NoReqBody jsonResponse opts
   pure (responseBody response)
@@ -1204,7 +1208,9 @@ getAttribution apiOpts revision ReportJson = fossaReq $ do
   jsonValue <- getAttributionJson apiOpts revision
   pure . decodeUtf8 $ Aeson.encode jsonValue
 getAttribution apiOpts ProjectRevision{..} format = fossaReq $ do
-  (baseUrl, opts) <- useApiOpts apiOpts
+  (baseUrl, baseOpts) <- useApiOpts apiOpts
+  -- Large reports can take over a minute to generate, so increase the timeout to 10 minutes
+  let opts = baseOpts <> responseTimeoutSeconds 600
 
   orgId <- organizationId <$> getOrganization apiOpts
   response <- req GET (attributionEndpoint baseUrl orgId (Locator "custom" projectName (Just projectRevision)) format) NoReqBody bsResponse opts
@@ -1696,8 +1702,8 @@ deleteReleaseGroup apiOpts releaseGroupId = fossaReq $ do
     context "Deleting release group" $
       req DELETE (deleteReleaseGroupURLEndpoint baseUrl $ toText releaseGroupId) NoReqBody ignoreResponse baseOpts
 
-releaseGroupReleaseURLEndpoint :: Url 'Https -> Text -> Text -> Url 'Https
-releaseGroupReleaseURLEndpoint baseUrl releaseGroupId releaseId = baseUrl /: "api" /: "project_group" /: releaseGroupId /: "release" /: releaseId
+specifiedReleaseGroupReleaseURLEndpoint :: Url 'Https -> Text -> Text -> Url 'Https
+specifiedReleaseGroupReleaseURLEndpoint baseUrl releaseGroupId releaseId = baseUrl /: "api" /: "project_group" /: releaseGroupId /: "release" /: releaseId
 
 deleteReleaseGroupRelease ::
   ( Has (Lift IO) sig m
@@ -1712,7 +1718,24 @@ deleteReleaseGroupRelease apiOpts releaseGroupId releaseId = fossaReq $ do
   (baseUrl, baseOpts) <- useApiOpts apiOpts
   void $
     context "Deleting release group release" $
-      req DELETE (releaseGroupReleaseURLEndpoint baseUrl (toText releaseGroupId) $ toText releaseId) NoReqBody ignoreResponse baseOpts
+      req DELETE (specifiedReleaseGroupReleaseURLEndpoint baseUrl (toText releaseGroupId) $ toText releaseId) NoReqBody ignoreResponse baseOpts
+
+updateReleaseGroupRelease ::
+  ( Has (Lift IO) sig m
+  , Has Diagnostics sig m
+  , Has Debug sig m
+  ) =>
+  ApiOpts ->
+  Int ->
+  Int ->
+  UpdateReleaseRequest ->
+  m ReleaseGroupRelease
+updateReleaseGroupRelease apiOpts releaseGroupId releaseId updateReq = fossaReq $ do
+  (baseUrl, baseOpts) <- useApiOpts apiOpts
+  resp <-
+    context "Updating release group release" $
+      req PUT (specifiedReleaseGroupReleaseURLEndpoint baseUrl (toText releaseGroupId) $ toText releaseId) (ReqBodyJson updateReq) jsonResponse baseOpts
+  pure (responseBody resp)
 
 releaseGroupURLEndpoint :: Url 'Https -> Url 'Https
 releaseGroupURLEndpoint baseUrl = baseUrl /: "api" /: "project_group"
@@ -1723,13 +1746,13 @@ createReleaseGroup ::
   , Has Debug sig m
   ) =>
   ApiOpts ->
-  CoreTypes.CreateReleaseGroupRequest ->
-  m CoreTypes.CreateReleaseGroupResponse
+  CreateReleaseGroupRequest ->
+  m CreateReleaseGroupResponse
 createReleaseGroup apiOpts rev = fossaReq $ do
   (baseUrl, baseOpts) <- useApiOpts apiOpts
   resp <-
     context "Creating release group" $
-      req POST (releaseGroupURLEndpoint baseUrl) (ReqBodyJson rev) jsonResponse baseOpts
+      req POST (releaseGroupURLEndpoint baseUrl) (ReqBodyJson createReleaseGroupReq) jsonResponse baseOpts
   pure (responseBody resp)
 
 getReleaseGroups ::
@@ -1746,8 +1769,8 @@ getReleaseGroups apiOpts = fossaReq $ do
       req GET (releaseGroupURLEndpoint baseUrl) NoReqBody jsonResponse baseOpts
   pure (responseBody resp)
 
-getReleaseGroupReleasesURLEndpoint :: Url 'Https -> Text -> Url 'Https
-getReleaseGroupReleasesURLEndpoint baseUrl releaseGroupId = baseUrl /: "api" /: "project_group" /: releaseGroupId /: "release"
+releaseGroupReleaseURLEndpoint :: Url 'Https -> Text -> Url 'Https
+releaseGroupReleaseURLEndpoint baseUrl releaseGroupId = baseUrl /: "api" /: "project_group" /: releaseGroupId /: "release"
 
 getReleaseGroupReleases ::
   ( Has (Lift IO) sig m
@@ -1761,10 +1784,10 @@ getReleaseGroupReleases apiOpts releaseGroupId = fossaReq $ do
   (baseUrl, baseOpts) <- useApiOpts apiOpts
   resp <-
     context "Retrieving release group releases" $
-      req GET (getReleaseGroupReleasesURLEndpoint baseUrl $ toText releaseGroupId) NoReqBody jsonResponse baseOpts
+      req GET (releaseGroupReleaseURLEndpoint baseUrl $ toText releaseGroupId) NoReqBody jsonResponse baseOpts
   pure (responseBody resp)
 
-updateReleaseGroupRelease ::
+createReleaseGroupRelease ::
   ( Has (Lift IO) sig m
   , Has Diagnostics sig m
   , Has Debug sig m
@@ -1772,13 +1795,13 @@ updateReleaseGroupRelease ::
   ApiOpts ->
   Int ->
   Int ->
-  CoreTypes.UpdateReleaseRequest ->
-  m CoreTypes.ReleaseGroupRelease
+  UpdateReleaseRequest ->
+  m ReleaseGroupRelease
 updateReleaseGroupRelease apiOpts releaseGroupId releaseId updateReq = fossaReq $ do
   (baseUrl, baseOpts) <- useApiOpts apiOpts
   resp <-
-    context "Updating release group release" $
-      req PUT (releaseGroupReleaseURLEndpoint baseUrl (toText releaseGroupId) $ toText releaseId) (ReqBodyJson updateReq) jsonResponse baseOpts
+    context "Creating release group release" $
+      req POST (releaseGroupReleaseURLEndpoint baseUrl $ toText releaseGroupId) (ReqBodyJson createReleaseReq) jsonResponse baseOpts
   pure (responseBody resp)
 
 updateProjectURLEndpoint :: Url 'Https -> Text -> Url 'Https